+++ CHANGELOG.md
@@ -44,12 +44,13 @@ NOTE: The brackets around the word "Unreleased" are required to pass the [CI tes
 that checks if we updated the changelog. This is a convention from the [keep a
 changelog format](https://keepachangelog.com/en/1.0.0/).  
 See https://github.com/dangoslen/changelog-enforcer.
---> 
+-->
 
 - [#2311](https://github.com/NibiruChain/nibiru/pull/2311) - refactor: use Go's built-in min and max functions to simplify logic
 - [#2314](https://github.com/NibiruChain/nibiru/pull/2314) - refactor(upgrades): add public keepers to upgrade handlers + DRY improvements
 - [#2315](https://github.com/NibiruChain/nibiru/pull/2315) - feat(upgrades): implement v2.5.0 upgrade handler that modifies the stNIBI ERC20 and Bank Coin metadata in place
 - [#2316](https://github.com/NibiruChain/nibiru/pull/2316) - feat(ux): add GET behavior to the Ethereum JSON-RPC endpoints for Nibiru so they return info instead of a blank page or error.
+- [#2324](https://github.com/NibiruChain/nibiru/pull/2324) - fix(evm): adjust the v2.5.0 upgrade handler to maintain the original stNIBI ERC20 contract's state.
 
 ## v2.4.0